consolidating ecs.js and lambda.js into one papertrail.js
The only significant difference in these two JS files is the
determination of the values for the `hostname` and `program`.

Replacing the duplication of the files with a single conditional block,
which populates those values depending on whether dealing with logs from
and ECS service or a Lambda.
